+++ source/cpp/core/ContextManager.h
@@ -1,8 +1,8 @@
 #ifndef CONTEXTMANAGER_H
 #define CONTEXTMANAGER_H
 
-#include "./CrossReferenceDispatcher.h"
-#include "./Context.h"
+#include "Context.h"
+#include "types.h"
 
 //TODO COMMENT !!!!!!!
 
@@ -14,22 +14,22 @@
  */
 struct ContextManager
 {
-	int_fast8_t current;			// Maximum of 256 contexts
+	uint8 current;			// Maximum of 256 contexts
 	std::vector<Context> contexts;
 
 	ContextManager()
 	{ current = 0; }
 
-	inline void setContext(int_fast8_t id)
+	inline void setContext(uint8 id)
 	{
-		if ((id >= 0) && (id < int_fast8_t(contexts.size())))
+		if ( id>=0 && id<contexts.size() )
 			current = id;
 		else throw std::exception();
 	}
 
-	inline void deleteContext(int_fast8_t id)
+	inline void deleteContext(uint8 id)
 	{
-		if (id >= 0 && id < int_fast8_t(contexts.size()))
+		if( id>=0 && id<contexts.size() )
 			contexts.erase(contexts.begin() + id);
 	}
