+++ source/cpp/core/SphereModel.h
@@ -10,19 +10,21 @@
 #define libsgl_SphereModel_h
 #include <cstdio>
 #include "../struct/Material.h"
+#include "../struct/Vertex.h"
+#include "Model.h"
 #include <vector>
 
 class SphereModel : public Model
 {
 public:
 	Vertex position;
-	float r;
+	uint16 r;
 	/**
 	 Constructor of Spehre model
 	 @param _g			drawing library
 	 @param _context	graphics context
 	 */
-	SphereModel( Chunk _context, Material _material, const float _x, const float _y, const float _z, const float _r) : Model( _context, _material)
+	SphereModel( Chunk _context, Material _material, const float _x, const float _y, const float _z, const uint16 _r ) : Model( _context, _material)
 	{
 		position.x	= _x;
 		position.y	= _y;
